pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "build-pdf"
version = "0.1.0"
description = "PDF Generation Library"
readme = "README.md"
authors = [
  { name = "Try3D" },
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
requires-python = ">=3.7"
dependencies = [
    "reportlab~=4.4.9",
    "pillow~=11.3.0",
]

[project.urls]
"Homepage" = "https://github.com/try3d/build-pdf"
"Bug Tracker" = "https://github.com/try3d/build-pdf/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["build_pdf*"]
